+++ Sources/KeyboardKit/_Essentials/Keyboard+Accent.swift
@@ -1,0 +1,18 @@
+//
+//  Keyboard+Accent.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2024-05-11.
+//  Copyright © 2024 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public extension Keyboard {
+    
+    /// This is a typealias for the ``Diacritic`` type.
+    ///
+    /// > Note: This typealias is meant to make it easier to
+    /// find the ``Diacritic`` feature.
+    typealias Accent = Diacritic
+}