+++ Sources/KeyboardKit/_Essentials/KeyboardController.swift
@@ -26,7 +26,10 @@ public protocol KeyboardController: AnyObject {
 
     /// Dismiss the keyboard.
     func dismissKeyboard()
-
+    
+    /// Insert the provided diacritic.
+    func insertDiacritic(_ diacritic: Keyboard.Diacritic)
+    
     /// Insert the provided text.
     func insertText(_ text: String)